feat: new initial connect account ux
